+++ packages/xrpl/HISTORY.md
@@ -5,6 +5,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ## Unreleased
 
 ### Added
+* Support for `Token Escrow` (XLS-85)
 * Support for `Single Asset Vault` (XLS-65)
 * Adds `XRPLNumber` amount type used in Vault transactions. This supports integer, decimal, or scientific notation strings.
 * Adds `ClawbackAmount` amount type used in transactions related to Clawback.